modularize codebase
